+++ NEWS.md
@@ -1,3 +1,11 @@
+# portion 0.1.2
+
+* Object attributes are preserved.
+
+* Added fallback `portion.default()` method.
+
+* Small bug fixes.
+
 # portion 0.1.1
 
 * Added methods for `character` and `logical`.